Added readonly variables for email creation
